feat: chuc nang crud service, thong ke danh sach booking, valid du lieu dau vao

+++ src/res/component/Room/RowTableRoomOfType.js
@@ -7,22 +7,24 @@ const RowTableRoomOfType = ({data,setRoom,setRoomsOfType}) => {
         setRoom(data)
     } 
     const handleDeleteRoom = async () => {
-        try{
-            setIsLoading(true)
-            const res = await fetch(apiUserService.baseURL+`/rooms/${data?.id}`,{
-                method:'DELETE'
-            })
-            setIsLoading(false)
-            if(res.ok){
-                const dataRes = await res.text()
-                if(dataRes === `Xóa phòng có id ${data.id} thành công!`){
-                    setRoomsOfType( rooms => rooms.filter( item => item.id !== data.id))
+        if(window.confirm("Bạn có muốn xóa phòng này ?")){
+            try{
+                setIsLoading(true)
+                const res = await fetch(apiUserService.baseURL+`/rooms/${data?.id}`,{
+                    method:'DELETE'
+                })
+                if(res.ok){
+                    const dataRes = await res.text()
+                    if(dataRes === `Xóa phòng có id ${data.id} thành công!`){
+                        setRoomsOfType( rooms => rooms.filter( item => item.id !== data.id))
+                    }
                 }
+                setIsLoading(false)
             }
-        }
-        catch(err){
-            setIsLoading(false)
-            console.log("loi xoa phong",err)
+            catch(err){
+                setIsLoading(false)
+                console.log("loi xoa phong",err)
+            }
         }
     }
     return <>
@@ -32,7 +34,11 @@ const RowTableRoomOfType = ({data,setRoom,setRoomsOfType}) => {
                 <td style={{textAlign:'center'}}>{data?.checkIn || "-"}</td>
                 <td style={{textAlign:'center'}}>{data?.checkOut || "-"}</td>
                 <td style={{textAlign:'center'}}>
-                    <button className="btn btn-danger" style={{marginRight:'10px'}} onClick={handleDeleteRoom}>Xóa</button>
+                    <button className="btn btn-danger" 
+                            style={{marginRight:'10px',cursor:(data?.status!=='trống')?'not-allowed':'pointer'}}
+                            onClick={handleDeleteRoom} disabled={(data?.status!=='trống')?true:false}>
+                            Xóa
+                    </button>
                     <button className="btn btn-info" onClick={handleIsClick}>Sửa</button>
                 </td>
             </tr>
